move all testing to one file
